Face Recognition Module/requirements.txt
# Face Recognition Module - Python Dependencies
# Install with: pip install -r requirements.txt

# Core Web Framework
fastapi
fastapi>=0.68.0
uvicorn>=0.15.0
flask

# Face Recognition and Deep Learning
insightface>=0.7.0
onnxruntime>=1.13.0
# Uncomment below for GPU support (CUDA)
# onnxruntime-gpu>=1.13.0

# Image Processing
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0

# Numerical Computing
numpy>=1.19.0

# Database
# SQLite3 is included with Python by default

# Security & Encryption
cryptography>=3.4.0

# Optional: Data Validation
pydantic>=1.8.0

# Optional: HTTP Client for API testing
requests>=2.26.0
pymongo
python-dotenv>=1.0.0
# Optional: Development
# Uncomment these for development purposes
# pytest>=6.2.0
# pytest-cov>=2.12.0
# black>=21.0  (Code formatter)
# flake8>=3.9.0  (Linter)
# mypy>=0.910  (Type checker)
certifi
